refactor: 수강 신청 관련 파일 폴더 정리
